backend/requirements.txt
absl-py==2.2.2
altgraph==0.17.4
anaconda==0.0.1.1
anyio==4.11.0
appdirs==1.4.4
apturl==0.5.2
asgiref==3.8.1
astunparse==1.6.3
attrs==21.2.0
beautifulsoup4==4.10.0
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
build==1.3.0
buildozer==1.5.0
CacheControl==0.14.3
cachetools==6.2.1
catfish==4.16.4
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
contourpy==1.3.2
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
Cython==3.1.4
dbus-python==1.2.18
defer==1.0.6
distlib==0.4.0
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.2.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
EditorConfig==0.12.2
et_xmlfile==2.0.0
evdev==1.4.0
exceptiongroup==1.3.0
filelock==3.19.1
firebase_admin==7.1.0
flatbuffers==25.2.10
fonttools==4.58.0
gast==0.6.0
google-api-core==2.28.1
google-auth==2.42.1
google-cloud-core==2.5.0
google-cloud-firestore==2.21.0
google-cloud-storage==3.4.1
google-crc32c==1.7.1
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.71.0
gpg==1.16.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.76.0
gyp==0.1
h11==0.16.0
h2==4.3.0
h5py==3.13.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.20.2
httpx==0.28.1
hyperframe==6.1.0
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
Jinja2==3.1.6
joblib==1.5.1
jsbeautifier==1.14.0
jsonschema==3.2.0
kazam==1.4.5
keras==3.9.2
keyring==23.5.0
kiwisolver==1.4.8
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==18.1.1
louis==3.20.0
lxml==4.8.0
Mako==1.1.3
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
menulibre==2.4.0
meson==0.61.2
ml_dtypes==0.5.1
more-itertools==8.10.0
msgpack==1.1.2
mugshot==0.4.3
mutagen==1.45.1
mysql-connector==2.2.9
mysql-connector-python==9.3.0
namex==0.0.9
natsort==8.0.2
netifaces==0.11.0
nodemon==0.1.1
numpy==2.1.3
oauthlib==3.2.0
olefile==0.46
onboard==1.4.1
opencv-python==4.12.0.88
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
pandas==2.2.3
parameterized==0.9.0
pexpect==4.8.0
pillow==10.4.0
platformdirs==4.4.0
proto-plus==1.26.1
protobuf==6.33.0
psutil==5.9.0
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.20.1
pycryptodomex==3.11.0
pycups==2.0.1
Pygments==2.19.1
PyGObject==3.42.1
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
PyJWT==2.10.1
pyparsing==2.4.7
pyproject_hooks==1.2.0
PyQt5==5.15.6
PyQt5-sip==12.9.1
PyQt6==6.4.2
pyqt6-plugins==6.4.2.2.3
PyQt6-Qt6==6.4.3
pyqt6-tools==6.4.2.3.3
PyQt6_sip==13.10.2
pyrsistent==0.18.1
python-apt===2.4.0-ubuntu4-zorin1
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
python-magic==0.4.24
pytz==2025.2
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.4.1
pyzmq==22.3.0
qrcode==8.2
qt6-applications==6.4.3.2.3
qt6-tools==6.4.3.1.3
reportlab==3.6.8
requests==2.25.1
rich==14.0.0
rsa==4.9.1
scikit-learn==1.7.0
scipy==1.15.3
scour==0.38.2
seaborn==0.13.2
SecretStorage==3.3.1
setproctitle==1.2.2
sh==1.14.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.3.1
SQLAlchemy==2.0.43
sqlparse==0.5.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.37.1
tensorflow_cpu==2.19.0
termcolor==3.1.0
texttable==1.6.4
threadpoolctl==3.6.0
toml==0.10.2
tomli==2.2.1
ttkbootstrap==1.14.4
typing_extensions==4.13.2
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
vboxapi==1.0
virtualenv==20.34.0
wadllib==1.3.6
webencodings==0.5.1
websocket-client==1.2.3
websockets==9.1
Werkzeug==3.1.3
wrapt==1.17.2
wxPython==4.0.7
xdg==5
xkit==0.0.0
yt-dlp==2025.4.30
zipp==1.0.0
zorin-appearance==5.3
zorin-exec-guard==1.5
